+++ source/Waypoint.cpp
@@ -1,7 +1,3 @@
-//
-// Created by bgd on 01.07.2024.
-//
-
 #include "../header/Waypoint.h"
 #include "../header/ResourcesManager.h"
 
@@ -11,12 +7,17 @@ Waypoint::Waypoint(sf::Vector2f position, const std::string& label) :
 {
     m_body.setFillColor(sf::Color::White);
     sf::FloatRect bounds = m_body.getGlobalBounds();
-    m_body.setOrigin(bounds.left + bounds.width / 2.f, bounds.top + bounds.height / 2.f);
+
+    m_body.setOrigin(static_cast<int>(bounds.left + bounds.width / 2.f),
+                     static_cast<int>(bounds.top + bounds.height / 2.f)); // cast to fix blurry text
+
     m_body.setPosition(static_cast<int>(position.x), static_cast<int>(position.y));
 
     bounds = m_label.getGlobalBounds();
-    m_label.setOrigin(bounds.left + bounds.width / 2.f, bounds.top + bounds.height / 2.f);
-    m_label.setPosition(static_cast<int>(position.x), static_cast<int>(position.y) - 8);
+    m_label.setOrigin(static_cast<int>(bounds.left + bounds.width / 2.f), // cast to fix blurry text
+                      static_cast<int>(bounds.top + bounds.height / 2.f));
+
+    m_label.setPosition(static_cast<int>(position.x), static_cast<int>(position.y - 8));
 }
 
 sf::Vector2f Waypoint::getPosition() const {
